chore: 🤖 make channel-server campitable for serverless
