+++ packages/channels/src/index.ts
@@ -1,5 +1,8 @@
 export * from "./channel";
 export * from "./session-manager";
+export * from "./db-session-manager";
 export * from "./channel-conversation-handler";
 export * from "./adapters/telegram-adapter";
 export * from "./adapters/telegram-types";
+export * from "./server/channel-server";
+export * from "./server/operations";
